requirements.txt
fastapi
uvicorn[standard]
langgraph
langchain-core
langchain-google-genai
pydantic
python-dotenv
